requirements.txt
Django==1.6.2
Scrapy==0.24.4
South==0.8.4
Twisted==14.0.2
argparse==1.2.1
cffi==0.8.6
cryptography==0.7.2
cssselect==0.9.1
dj-database-url==0.3.0
dj-static==0.0.5
django-bootstrap3==4.2.0
django-extensions==1.3.3
django-jquery==1.9.1
django-tastypie==0.11.0
enum34==1.0.4
flake8==2.1.0
gunicorn==18.0
ipython==2.0.0
lxml==3.3.4
mccabe==0.2.1
newrelic==2.18.1.15
pep8==1.5.6
psycopg2==2.5.2
pyOpenSSL==0.14
pyasn1==0.1.7
pycparser==2.10
pyflakes==0.8.1
pystache==0.5.3
python-dateutil==2.2
python-mimeparse==0.1.4
queuelib==1.2.2
requests==2.2.1
six==1.6.1
static==1.0.2
w3lib==1.11.0
wsgiref==0.1.2
zope.interface==4.1.2
